+++ src/main/java/uk/gov/hmcts/reform/migration/auth/AuthTokenGeneratorConfiguration.java
@@ -1,4 +1,4 @@
-package uk.gov.hmcts.reform.fpl.ccddatamigration.auth;
+package uk.gov.hmcts.reform.migration.auth;
 
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Bean;
