+++ src/main/resources/application.properties
@@ -19,7 +19,7 @@ core_case_data.api.url=http://localhost:4452
 
 logging.level.root=ERROR
 logging.level.uk.gov.hmcts.reform=INFO
-logging.level.uk.gov.hmcts.reform.fpl=${log.level:INFO}
+logging.level.uk.gov.hmcts.reform.migration=${log.level:INFO}
 
 feign.client.config.default.connectTimeout=60000
 feign.client.config.default.readTimeout=60000
