+++ src/test/java/uk/gov/hmcts/reform/migration/auth/AuthUtilTest.java
@@ -1,4 +1,4 @@
-package uk.gov.hmcts.reform.fpl.ccddatamigration.auth;
+package uk.gov.hmcts.reform.migration.auth;
 
 import org.junit.Test;
